requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
dnspython==2.3.0
EasyProcess==1.1
entrypoint2==1.1
exceptiongroup==1.1.1
fastapi==0.99.0
Flask==2.3.2
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
numpy==1.25.0
outcome==1.2.0
pandas==2.0.3
Pillow==10.0.0
playwright==1.35.0
py==1.11.0
PyAutoGUI==0.9.54
pydantic==1.10.10
pydantic_core==2.0.1
pyee==9.0.4
PyGetWindow==0.0.9
PyJWT==2.7.0
pymongo==4.4.0
PyMsgBox==1.0.9
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
pywhatkit==5.4
requests==2.31.0
retry==0.9.2
rubicon-objc==0.4.6
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.27.0
trio==0.22.0
trio-websocket==0.10.3
twilio==8.4.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
Werkzeug==2.3.6
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.9.2
